Some useful functions were available.
Now FKConsole will save logs to disk automaticly, and you can set shouldSaveLogsToDisk to false to disable it.
Font in FKConsole can be customized now.
A UITextView is used for displaying logs, so you can choose log text and copy it with a long press. (UITableView is removed)
FKConsole will handle uncaught exception and output logs to disk.